index::inverted_index::create
Module sort

Modules

external_sort

intermediate_rw 🔒

Intermediate codec for external sorting.

merge_stream 🔒

Structs

SortOutput

Output of a sorting operation, encapsulating a bitmap for null values and a stream of sorted items

Traits

Sorter

Handles data sorting, supporting incremental input and retrieval of sorted output

Type Aliases

SortedStream

A stream of sorted values along with their associated bitmap