Trait index::inverted_index::create::sort::Sorter

pub trait Sorter: Send {
    // Required methods
    fn push_n<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        value: Option<BytesRef<'life1>>,
        n: usize
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn output<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<SortOutput>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;

    // Provided method
    fn push<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        value: Option<BytesRef<'life1>>
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
}

Handles data sorting, supporting incremental input and retrieval of sorted output

Required Methods

fn push_n<'life0, 'life1, 'async_trait>( &'life0 mut self, value: Option<BytesRef<'life1>>, n: usize ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Pushing n identical non-null or null values into the sorter. Should be equivalent to calling push n times

fn output<'life0, 'async_trait>( &'life0 mut self ) -> Pin<Box<dyn Future<Output = Result<SortOutput>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Completes the sorting process and returns the sorted data

Provided Methods

fn push<'life0, 'life1, 'async_trait>( &'life0 mut self, value: Option<BytesRef<'life1>> ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Inputs a non-null or null value into the sorter. Should be equivalent to calling push_n with n = 1