index::inverted_index::create

Trait InvertedIndexCreator

pub trait InvertedIndexCreator: Send {
    // Required methods
    fn push_with_name_n<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self,
        index_name: &'life1 str,
        value: Option<BytesRef<'life2>>,
        n: usize,
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn finish<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        writer: &'life1 mut dyn InvertedIndexWriter,
        bitmap_type: BitmapType,
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;

    // Provided method
    fn push_with_name<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self,
        index_name: &'life1 str,
        value: Option<BytesRef<'life2>>,
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait { ... }
}

InvertedIndexCreator provides functionality to construct an inverted index

Required Methods

fn push_with_name_n<'life0, 'life1, 'life2, 'async_trait>( &'life0 mut self, index_name: &'life1 str, value: Option<BytesRef<'life2>>, n: usize, ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Adds n identical values to the named index. None values represent absence of data (null)

index_name: Identifier for the index being built

value: The data to be indexed, or None for a null entry

It should be equivalent to calling push_with_name n times

fn finish<'life0, 'life1, 'async_trait>( &'life0 mut self, writer: &'life1 mut dyn InvertedIndexWriter, bitmap_type: BitmapType, ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Finalizes the index creation process, ensuring all data is properly indexed and stored in the provided writer

Provided Methods

fn push_with_name<'life0, 'life1, 'life2, 'async_trait>( &'life0 mut self, index_name: &'life1 str, value: Option<BytesRef<'life2>>, ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Adds a value to the named index. A None value represents an absence of data (null)

index_name: Identifier for the index being built

value: The data to be indexed, or None for a null entry

It should be equivalent to calling push_with_name_n with n = 1