index::inverted_index::format::reader

Trait InvertedIndexReader

pub trait InvertedIndexReader: Send + Sync {
    // Required methods
    fn range_read<'life0, 'async_trait>(
        &'life0 self,
        offset: u64,
        size: u32,
    ) -> Pin<Box<dyn Future<Output = Result<Vec<u8>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn metadata<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<Arc<InvertedIndexMetas>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;

    // Provided methods
    fn read_vec<'life0, 'life1, 'async_trait>(
        &'life0 self,
        ranges: &'life1 [Range<u64>],
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Bytes>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn fst<'life0, 'async_trait>(
        &'life0 self,
        offset: u64,
        size: u32,
    ) -> Pin<Box<dyn Future<Output = Result<FstMap>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }
    fn fst_vec<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        ranges: &'life1 [Range<u64>],
    ) -> Pin<Box<dyn Future<Output = Result<Vec<FstMap>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn bitmap<'life0, 'async_trait>(
        &'life0 self,
        offset: u64,
        size: u32,
        bitmap_type: BitmapType,
    ) -> Pin<Box<dyn Future<Output = Result<Bitmap>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }
    fn bitmap_deque<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        ranges: &'life1 [(Range<u64>, BitmapType)],
    ) -> Pin<Box<dyn Future<Output = Result<VecDeque<Bitmap>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
}

InvertedIndexReader defines an asynchronous reader of inverted index data

Required Methods

fn range_read<'life0, 'async_trait>( &'life0 self, offset: u64, size: u32, ) -> Pin<Box<dyn Future<Output = Result<Vec<u8>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Seeks to given offset and reads data with exact size as provided.

fn metadata<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<Arc<InvertedIndexMetas>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Retrieves metadata of all inverted indices stored within the blob.

Provided Methods

fn read_vec<'life0, 'life1, 'async_trait>( &'life0 self, ranges: &'life1 [Range<u64>], ) -> Pin<Box<dyn Future<Output = Result<Vec<Bytes>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Reads the bytes in the given ranges.

fn fst<'life0, 'async_trait>( &'life0 self, offset: u64, size: u32, ) -> Pin<Box<dyn Future<Output = Result<FstMap>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Retrieves the finite state transducer (FST) map from the given offset and size.

fn fst_vec<'life0, 'life1, 'async_trait>( &'life0 mut self, ranges: &'life1 [Range<u64>], ) -> Pin<Box<dyn Future<Output = Result<Vec<FstMap>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Retrieves the multiple finite state transducer (FST) maps from the given ranges.

fn bitmap<'life0, 'async_trait>( &'life0 self, offset: u64, size: u32, bitmap_type: BitmapType, ) -> Pin<Box<dyn Future<Output = Result<Bitmap>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Retrieves the bitmap from the given offset and size.

fn bitmap_deque<'life0, 'life1, 'async_trait>( &'life0 mut self, ranges: &'life1 [(Range<u64>, BitmapType)], ) -> Pin<Box<dyn Future<Output = Result<VecDeque<Bitmap>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Retrieves the multiple bitmaps from the given ranges.