index::inverted_index::format
Module writer

Modules

__mock_MockInvertedIndexWriter

__mock_MockInvertedIndexWriter_InvertedIndexWriter

blob 🔒

single 🔒

Structs

InvertedIndexBlobWriter

InvertedIndexBlobWriter, implemented InvertedIndexWriter, manages writing of an inverted index to a blob storage.

MockInvertedIndexWriter

Trait for writing inverted index data to underlying storage.

MockInvertedIndexWriter_InvertedIndexWriter 🔒

Traits

InvertedIndexWriter

Trait for writing inverted index data to underlying storage.

Type Aliases

ValueStream