index::inverted_index::search
Module fst_apply

Modules

__mock_MockFstApplier

__mock_MockFstApplier_FstApplier

intersection_apply 🔒

keys_apply 🔒

Structs

IntersectionFstApplier

IntersectionFstApplier applies intersection operations on an FstMap using specified ranges and regex patterns.

KeysFstApplier

KeysFstApplier is responsible for applying a search using a set of predefined keys against an FstMap to fetch associated values.

MockFstApplier

A trait for objects that can process a finite state transducer (FstMap) and return associated values.

MockFstApplier_FstApplier 🔒

Traits

FstApplier

A trait for objects that can process a finite state transducer (FstMap) and return associated values.