index::inverted_index::search

Module index_apply

Modules

__mock_MockIndexApplier

__mock_MockIndexApplier_IndexApplier

predicates_apply 🔒

Structs

ApplyOutput

The output of an apply operation.

MockIndexApplier

A trait for processing and transforming indices obtained from an inverted index.

MockIndexApplier_IndexApplier 🔒

PredicatesIndexApplier

PredicatesIndexApplier contains a collection of FstAppliers, each associated with an index name, to process and filter index data based on compiled predicates.

SearchContext

A context for searching the inverted index.

Enums

IndexNotFoundStrategy

Defines the behavior of an applier when the index is not found.

Traits

IndexApplier

A trait for processing and transforming indices obtained from an inverted index.