index::inverted_index::search::index_apply

Trait IndexApplier

pub trait IndexApplier: Send + Sync {
    // Required methods
    fn apply<'a, 'life0, 'life1, 'async_trait>(
        &'life0 self,
        context: SearchContext,
        reader: &'life1 mut (dyn InvertedIndexReader + 'a),
    ) -> Pin<Box<dyn Future<Output = Result<ApplyOutput>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'a: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn memory_usage(&self) -> usize;
}

A trait for processing and transforming indices obtained from an inverted index.

Applier instances are reusable and work with various InvertedIndexReader instances, avoiding repeated compilation of fixed predicates such as regex patterns.

Required Methods

fn apply<'a, 'life0, 'life1, 'async_trait>( &'life0 self, context: SearchContext, reader: &'life1 mut (dyn InvertedIndexReader + 'a), ) -> Pin<Box<dyn Future<Output = Result<ApplyOutput>> + Send + 'async_trait>>
where Self: 'async_trait, 'a: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Applies the predefined predicates to the data read by the given index reader, returning a list of relevant indices (e.g., post IDs, group IDs, row IDs).

fn memory_usage(&self) -> usize

Returns the memory usage of the applier.