Struct index::Index

pub struct Index<K, V, S = IndexHasherBuilder> { /* fields omitted */ }

Simple implementation of a hash table using safe-rust.

The collisions are resolved through open adressing with quadratic probing (although it is possible to use linear probing or other types when specifying parameters).

Example

use index::Index;
 
let mut index = Index::new();
 
assert_eq!(index.len(), 0);
assert_eq!(index.capacity(), 1);
 
index.insert("salutation", "Hello, world!");
index.insert("ferris", "https://www.rustacean.net/more-crabby-things/dancing-ferris.gif");
index.insert("did you know ?", "Rust is kinda cool guys !");
index.insert("key", "value");
 
println!("{}", index.get("salutation").unwrap());
 
assert_eq!(index.len(), 4);
assert_eq!(index.capacity(), 8);

Methods

impl<K, V> Index<K, V, IndexHasherBuilder> where
K: Hash + Eq,

pub fn new() -> Index<K, V, IndexHasherBuilder>

Creates an empty Index with default initial capacity and default parameters.

Example

use index::Index;
 
let mut index: Index<String, Vec<i32>> = Index::new();

pub fn with_capacity(capacity: usize) -> Index<K, V, IndexHasherBuilder>

Creates an empty Index with specified capacity and default parameters.

Example

use index::Index;
 
let mut index: Index<String, Vec<i32>> = Index::with_capacity(1312);

impl<K, V, S> Index<K, V, S>

pub fn max_load(&self) -> f64

Returns the maximum load factor accepted before the table is resized.

Example

use index::Index;
 
let mut index: Index<String, Vec<i32>> = Index::new();
 
assert_eq!(index.max_load(), 0.7); // default max load

pub fn growth_policy(&self) -> f64

Returns the ratio by which the table's capacity is grown.

Example

use index::Index;
 
let mut index: Index<String, Vec<i32>> = Index::new();
 
assert_eq!(index.growth_policy(), 2.0); // default growth policy

pub fn hasher(&self) -> &S

Returns a reference to the hasher builder used in the Index.

Example

use index::Index;
use index::hash::IndexHasherBuilder;
use std::any::{Any, TypeId};
 
let mut index: Index<String, Vec<i32>> = Index::new();
 
assert_eq!(index.hasher().type_id(), TypeId::of::<IndexHasherBuilder>()) // default hasher builder

pub fn probe(&self) -> fn(_: usize, _: usize) -> usize

Returns the probing function pointer of the Index.

Example

use index::Index;
 
let mut index: Index<String, Vec<i32>> = Index::new();
 
let p = |h: usize, i: usize| h + i + i*i; // default prober
 
assert_eq!((index.probe())(45, 2), p(45, 2));

pub fn capacity(&self) -> usize

Returns the capacity of the Index.

Example

use index::Index;
 
let mut index: Index<&str, &str> = Index::with_capacity(6);
 
assert_eq!(index.len(), 0);
assert_eq!(index.capacity(), 6);

pub fn len(&self) -> usize

Returns the number of elements in the Index.

Example

use index::Index;
 
let mut index: Index<&str, i32> = Index::with_capacity(6);
 
index.insert("one", 1);
index.insert("two", 2);
index.insert("three", 3);
 
assert_eq!(index.len(), 3);
assert_eq!(index.capacity(), 6);

pub fn is_empty(&self) -> bool

Returns true if the Index contains no elements.

Example

use index::Index;
 
let mut index: Index<&str, &str> = Index::with_capacity(10);
 
assert!(index.is_empty());

pub fn load(&self) -> f64

Returns the current load factor of the Index.

Example

use index::Index;
 
let mut index: Index<&str, i32> = Index::with_capacity(6);
 
index.insert("one", 1);
index.insert("two", 2);
index.insert("three", 3);
 
assert_eq!(index.load(), 0.5);

pub fn clear(&mut self)

Clear the Index, replacing all entries with empty buckets.

Example

use index::Index;
use index::Bucket;
 
let mut index: Index<&str, i32> = Index::with_capacity(6);
 
index.insert("one", 1);
index.insert("two", 2);
index.insert("three", 3);
 
index.clear();
 
assert!(index.get("two").is_none());
assert_eq!(index.len(), 0);
 

Important traits for Keys<'a, K, V>
impl<'a, K, V> Iterator for Keys<'a, K, V> type Item = Ref<'a, K>;

pub fn keys(&self) -> Keys<K, V>

Returns an iterator over the keys of the Index. The iterator's associated type is Ref<'a, K>.

Example

use index::Index;
 
let mut index = Index::with_capacity(10);
 
index.insert("salutation", "Hello, world!");
index.insert("ferris", "https://www.rustacean.net/more-crabby-things/dancing-ferris.gif");
index.insert("did you know ?", "Rust is kinda cool !");
 
for key in index.keys() {
    println!("key: {:?}", key);
}
 
assert_eq!(index.len(), index.keys().count());

Important traits for Values<'a, K, V>
impl<'a, K, V> Iterator for Values<'a, K, V> type Item = Ref<'a, V>;

pub fn values(&self) -> Values<K, V>

Returns an iterator over the values of the Index. The iterator's associated type is Ref<'a, V>.

Example

use index::Index;
 
let mut index = Index::with_capacity(10);
 
index.insert("salutation", "Hello, world!");
index.insert("ferris", "https://www.rustacean.net/more-crabby-things/dancing-ferris.gif");
index.insert("did you know ?", "Rust is kinda cool !");
 
for value in index.values() {
    println!("value: {:?}", value);
}
 
assert_eq!(index.len(), index.values().count());

Important traits for ValuesMut<'a, K, V>
impl<'a, K, V> Iterator for ValuesMut<'a, K, V> type Item = RefMut<'a, V>;

pub fn values_mut(&self) -> ValuesMut<K, V>

Returns a mutable iterator over the values of the Index. The iterator's associated type is RefMut<'a, V>.

Example

use index::Index;
 
let mut index = Index::with_capacity(10);
 
index.insert("salutation", "Hello, world!");
index.insert("ferris", "https://www.rustacean.net/more-crabby-things/dancing-ferris.gif");
index.insert("did you know ?", "Rust is kinda cool !");
 
for mut value in index.values_mut() {
    *value = "overwritten!";
}
 
assert_eq!(*index.get("ferris").unwrap(), "overwritten!");
 

Important traits for Iter<'a, K, V>
impl<'a, K, V> Iterator for Iter<'a, K, V> type Item = Ref<'a, (K, V)>;

pub fn iter(&self) -> Iter<K, V>

Return an iterator over the key-value pairs of the Index. The iterator's associated type is Ref<'a, (K, V)>.

Example

use index::Index;
 
let mut index = Index::with_capacity(10);
 
index.insert("salutation", "Hello, world!");
index.insert("ferris", "https://www.rustacean.net/more-crabby-things/dancing-ferris.gif");
index.insert("did you know ?", "Rust is kinda cool !");
 
for entry in index.iter() {
    println!("key: {:?} => value: {:?}", entry.0, entry.1);
}
 
assert_eq!(index.len(), index.iter().count());

Important traits for IterMut<'a, K, V>
impl<'a, K, V> Iterator for IterMut<'a, K, V> type Item = RefMut<'a, (K, V)>;

pub fn iter_mut(&self) -> IterMut<K, V>

Return a mutable iterator over the key-value pairs of the Index. The iterator's associated type is RefMut<'a, (K, V)>.

Example

use index::Index;
 
let mut index = Index::with_capacity(10);
 
index.insert("salutation", "Hello, world!");
index.insert("ferris", "https://www.rustacean.net/more-crabby-things/dancing-ferris.gif");
index.insert("did you know ?", "Rust is kinda cool !");
 
for mut entry in index.iter_mut() {
    entry.1 = entry.0.clone();
}
 
assert_eq!(*index.get("ferris").unwrap(), "ferris");

Important traits for Drain<'a, K, V>
impl<'a, K, V> Iterator for Drain<'a, K, V> type Item = (K, V);

pub fn drain(&mut self) -> Drain<K, V>

Returns iterator taking ownership and moving out the key-value pairs of the Index.

Example

use index::Index;
 
let mut index = Index::with_capacity(10);
 
index.insert("salutation", "Hello, world!");
index.insert("ferris", "https://www.rustacean.net/more-crabby-things/dancing-ferris.gif");
index.insert("did you know ?", "Rust is kinda cool !");
 
let v: Vec<(&str, &str)> = index.drain().collect();
 
assert_eq!(index.len(), 0);
assert_eq!(v.len(), 3);
assert!(v.contains(&("salutation", "Hello, world!")));

impl<K, V, S> Index<K, V, S> where
K: Hash + Eq,
S: BuildHasher + Clone,

pub fn with_capacity_and_parameters(
capacity: usize,
params: Parameters<S>
) -> Index<K, V, S>

Creates an empty Index with specified capacity and parameters.

See Parameters for details.

Example

use std::collections::hash_map::RandomState;
use index::{Index, Parameters};
 
let params = Parameters {
    max_load: 0.7,
    growth_policy: 2.0,
    hasher_builder: RandomState::new(),
    probe: |hash, i| (hash as f64 + (i as f64 / 2.0) + ((i*i) as f64 / 2.0)) as usize,
};
 
let mut index = Index::with_capacity_and_parameters(10, params);
 
index.insert("key", "value");

pub fn insert(&mut self, key: K, value: V) -> Bucket<K, V>

Inserts key-value pair in the Index.

If it encounters an occupied bucket with the same key, it will replace the entry according to the new value and return the old bucket.

The function also verifies before anything else that the load factor is lesser than the maximum accepted load, if not it will grow the Index before proceeding to the insertion.

If the lookup returns no valid result, the insertion is considered impossible and the function will grow the Index and retry to insert the pair.

Example

use index::Index;
 
let mut index = Index::with_capacity(2);
 
index.insert("key", "value");
 
assert_eq!(*index.get("key").unwrap(), "value");
 
index.insert("key", "new value");
 
assert_eq!(*index.get("key").unwrap(), "new value");
 
assert_eq!(index.len(), 1);
assert_eq!(index.capacity(), 2);
 
index.insert("salutation", "Hello, world!");
index.insert("ferris", "https://www.rustacean.net/more-crabby-things/dancing-ferris.gif");
index.insert("did you know ?", "Rust is kinda cool guys !");
 
assert_eq!(index.len(), 4);
assert_eq!(index.capacity(), 8);

pub fn get<Q: ?Sized>(&self, key: &Q) -> Option<Ref<V>> where
K: Borrow<Q>,
Q: Hash + Eq,

Returns a reference to the value associated with the specified key if the lookup found a match, else it returns None.

Example

use index::Index;
 
let mut index = Index::with_capacity(10);
 
index.insert("salutation", "Hello, world!");
index.insert("ferris", "https://www.rustacean.net/more-crabby-things/dancing-ferris.gif");
index.insert("did you know ?", "Rust is kinda cool !");
 
assert_eq!(*index.get("salutation").unwrap(), "Hello, world!");

pub fn get_mut<Q: ?Sized>(&self, key: &Q) -> Option<RefMut<V>> where
K: Borrow<Q>,
Q: Hash + Eq,

Returns a mutable reference to the value associated with the specified key if the lookup found a match, else it returns None.

Example

use index::Index;
 
let mut index = Index::with_capacity(10);
 
index.insert("salutation", "Hello, world!");
index.insert("ferris", "https://www.rustacean.net/more-crabby-things/dancing-ferris.gif");
index.insert("did you know ?", "Rust is kinda cool !");
 
*index.get_mut("salutation").unwrap() = "Hello, rust!";
 
assert_eq!(*index.get("salutation").unwrap(), "Hello, rust!");

pub fn get_pair<Q: ?Sized>(&self, key: &Q) -> Option<Ref<(K, V)>> where
K: Borrow<Q>,
Q: Hash + Eq,

Returns a reference to the key-value pair associated with the specified key if the lookup found a match, else it returns None.

Example

use index::Index;
 
let mut index = Index::with_capacity(10);
 
index.insert("salutation", "Hello, world!");
index.insert("ferris", "https://www.rustacean.net/more-crabby-things/dancing-ferris.gif");
index.insert("did you know ?", "Rust is kinda cool !");
 
assert_eq!(*index.get_pair("did you know ?").unwrap(), ("did you know ?", "Rust is kinda cool !"));