index

Module target

Structs

EmptySnafu 🔒

SNAFU context selector for the TargetKeyError::Empty variant

InvalidCharactersSnafu 🔒

SNAFU context selector for the TargetKeyError::InvalidCharacters variant

InvalidColumnIdSnafu 🔒

SNAFU context selector for the TargetKeyError::InvalidColumnId variant

Enums

IndexTarget

Describes an index target. Column ids are the only supported variant for now.

TargetKeyError

Errors that can occur when working with index target keys.

Functions

validate_column_key 🔒