Trait index::inverted_index::format::reader::InvertedIndexReader

pub trait InvertedIndexReader: Send {
    // Required methods
    fn read_all<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        dest: &'life1 mut Vec<u8>
    ) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn seek_read<'life0, 'async_trait>(
        &'life0 mut self,
        offset: u64,
        size: u32
    ) -> Pin<Box<dyn Future<Output = Result<Vec<u8>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn metadata<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<Arc<InvertedIndexMetas>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;

    // Provided methods
    fn fst<'life0, 'async_trait>(
        &'life0 mut self,
        offset: u64,
        size: u32
    ) -> Pin<Box<dyn Future<Output = Result<FstMap>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }
    fn bitmap<'life0, 'async_trait>(
        &'life0 mut self,
        offset: u64,
        size: u32
    ) -> Pin<Box<dyn Future<Output = Result<BitVec>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }
}

InvertedIndexReader defines an asynchronous reader of inverted index data

Required Methods

fn read_all<'life0, 'life1, 'async_trait>( &'life0 mut self, dest: &'life1 mut Vec<u8> ) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Reads all data to dest.

fn seek_read<'life0, 'async_trait>( &'life0 mut self, offset: u64, size: u32 ) -> Pin<Box<dyn Future<Output = Result<Vec<u8>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Seeks to given offset and reads data with exact size as provided.

fn metadata<'life0, 'async_trait>( &'life0 mut self ) -> Pin<Box<dyn Future<Output = Result<Arc<InvertedIndexMetas>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Retrieves metadata of all inverted indices stored within the blob.

Provided Methods

fn fst<'life0, 'async_trait>( &'life0 mut self, offset: u64, size: u32 ) -> Pin<Box<dyn Future<Output = Result<FstMap>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Retrieves the finite state transducer (FST) map from the given offset and size.

fn bitmap<'life0, 'async_trait>( &'life0 mut self, offset: u64, size: u32 ) -> Pin<Box<dyn Future<Output = Result<BitVec>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Retrieves the bitmap from the given offset and size.